Fetch user data in dashboard header since that's where it really belongs
