+++ src/layouts/dashboardLayout.js
@@ -1,13 +1,14 @@
 import React from 'react'
-import ReactLoading from 'react-loading'
 import DashboardHeader from '../components/dashboardHeader/dashboardHeader'
 import styles from './dashboardLayout.module.css'
 
-const DashboardLayout = ({ userData, children }) => (
-  <div className={styles.root}>
-    {children}
-    {!!userData ? <DashboardHeader data={userData} /> : <ReactLoading type='spinningBubbles' color='#ccc' />}
-  </div>
-)
+const DashboardLayout = ({ children }) => {
+  return(
+    <div className={styles.root}>
+      {children}
+      <DashboardHeader />
+    </div>
+  )
+}
 
 export default DashboardLayout
